[FEAT] Card / List 객체 구성 변경
- Card 클래스가 별도로 있고, List 가 내부 Card 클래스를 가지고 있도록 변경
